pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite==0.21.0",
    "alembic==1.15.1",
    "asyncpg==0.30.0",
    "fastapi==0.115.11",
    "orjson==3.10.15",
    "psycopg==3.2.5",
    "psycopg-c==3.2.5; sys_platform != 'win32'",
    "pydantic==2.10.6",
    "pydantic-settings==2.8.1",
    "pyjwt==2.10.1",
    "rich==13.9.4",
    "sqlmodel==0.0.23",
    "uvicorn[standard]==0.34.0",
    "uvloop==0.21.0; sys_platform != 'win32'",
]

[dependency-groups]
dev = [
    "pip-check==2.9",
    "poethepoet==0.33.0",
    "pre-commit==4.1.0",
    "psycopg[binary]==3.2.5",
    "types-jwt==0.1.3",
]

[tool.poe.tasks]
_git = "git add ."
_lint = "pre-commit run --all-files"

lint = ["_git", "_lint"]
check = "uv run pip-check -H"
run = "uv run src/server.py"
migrations = "uv run alembic revision --autogenerate"

[tool.uv]
required-version = ">=0.5.29"
dependency-metadata = [
    { name = "psycopg-c", version = "3.2.5", python = ">=3.8", requires = ["psycopg==3.2.5"] }
]

[tool.typos.files]
extend-exclude = ["**/migrations/versions"]

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["**/migrations/versions"]
strictListInference = true
strictDictionaryInference = true
strictSetInference = true
deprecateTypingAliases = true
typeCheckingMode = "strict"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
extend-select = ["F", "UP", "B", "SIM", "I"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
